Enable loading 10M records in Mongo
